mcp/requirements.txt
# MCP Server Framework
# Note: The official MCP Python SDK package name may vary
# Install with: pip install mcp
# Or use: pip install git+https://github.com/modelcontextprotocol/python-sdk.git
mcp

# Document Parsing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
markdown>=3.5.0

# Configuration and Utilities
PyYAML>=6.0.0

# HTTP Server (for remote access)
uvicorn>=0.24.0
fastapi>=0.104.0
